backend/pyproject.toml
[project]
name = "realtrust-api"
version = "0.1.0"
description = "realtrust ai backend API — FastAPI + PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pgvector>=0.2.0",
    "geoalchemy2>=0.14.0",
    "boto3>=1.34.0",
    "minio>=7.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21.0", "httpx>=0.25.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/realtrust_api"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "psycopg2-binary>=2.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
pythonpath = ["src"]
